django_core/requirements.txt
Django>=5.1.0,<5.2.0
psycopg2-binary>=2.9.6,<2.10
pytest-django>=4.11.1,<4.12.0
djangorestframework>=3.16.0,<3.17.0